requirements.txt
openai-whisper
sounddevice
numpy
torch
transformers
ctransformers
tqdm
pytest
